security-release-process: document maximal representation state
We are currently overlimit using the old rules and would continue to be
if we added two new members. Given the committee has quite a bit of
work, and is working on onboarding new members right now I think this
updated policy reflects the honest state of where we are.