README: add Luke Hinds to Security Committee
I want to nominate Luke Hinds to join the Product Security Committee as
a full member. He has been an associate member since June 26th 2019
and in that role over the last three months he has helped on a number of
tasks including:

- Coordinating CVE-2019-11251 announcement
- Writing Kubernetes Security Committee on-boarding guide

Our process for accepting this nomination is lazy consensus and I will
close the round of consensus on Oct 18th, 2019 to leave time for any
comments or question. If the nomination is accepted Luke will take on
the full responsibilities of a Security Committee member and take on a
weekly rotation for security@kubernetes.io on-call rotation.